+++ apps/frontend/src/libs/types/types.ts
@@ -6,6 +6,7 @@ export {
 	type APIErrorResponse,
 	type APIResponse,
 	type APIValidationErrorDetail,
+	type Coordinates,
 	type ValidationSchema,
 	type ValueOf,
 } from "@smartscapes/shared";