+++ apps/frontend/src/modules/points-of-interest/slices/actions.ts
@@ -6,29 +6,30 @@ import {
 	type PointsOfInterestResponseDto,
 } from "~/modules/points-of-interest/points-of-interest.js";
 
-import { name as sliceName } from "./points-of-interest.slice.js";
+import { name as poiDetailsSliceName } from "./points-of-interest-details.slice.js";
+import { name as poiSliceName } from "./points-of-interest.slice.js";
+
+const getById = createAsyncThunk<
+	APIResponse<PointsOfInterestResponseDto>,
+	number,
+	AsyncThunkConfig
+>(`${poiDetailsSliceName}/load-by-id`, (id, { extra }) => {
+	const { pointOfInterestApi } = extra;
+
+	return pointOfInterestApi.getById(id);
+});
 
 const create = createAsyncThunk<
 	APIResponse<PointsOfInterestResponseDto>,
 	PointsOfInterestRequestDto,
 	AsyncThunkConfig
->(`${sliceName}/create`, async (payload, { extra }) => {
+>(`${poiSliceName}/create`, async (payload, { extra }) => {
 	const { pointOfInterestApi, toastNotifier } = extra;
 
 	const pointOfInterest = await pointOfInterestApi.create(payload);
 	toastNotifier.showSuccess("Point of interest created successfully");
 
 	return pointOfInterest;
-});
-
-const getById = createAsyncThunk<
-	APIResponse<PointsOfInterestResponseDto>,
-	number,
-	AsyncThunkConfig
->(`${sliceName}/get-by-id`, async (id, { extra }) => {
-	const { pointOfInterestApi } = extra;
-
-	return await pointOfInterestApi.getById(id);
 });
 
 export { create, getById };