+++ apps/frontend/src/modules/points-of-interest/slices/points-of-interest.ts
@@ -1,11 +1,14 @@
 import { create, getById } from "./actions.js";
+import { actions as poiDetailsActions } from "./points-of-interest-details.slice.js";
 import { actions } from "./points-of-interest.slice.js";
 
 const allActions = {
 	...actions,
+	...poiDetailsActions,
 	create,
 	getById,
 };
 
 export { allActions as actions };
-export { reducer } from "./points-of-interest.slice.js";
+export { reducer as pointOfInterestDetailsReducer } from "./points-of-interest-details.slice.js";
+export { reducer as pointsOfInterestReducer } from "./points-of-interest.slice.js";