+++ apps/frontend/src/pages/route-description/libs/components/components.ts
@@ -1,1 +1,0 @@
-export { ImageGallery } from "./image-gallery/image-gallery.js";