+++ packages/shared/src/modules/routes/libs/types/route-get-all-item-response-dto.type.ts
@@ -1,4 +1,10 @@
+import { type LineStringGeometry } from "../../../../libs/types/types.js";
+
 type RouteGetAllItemResponseDto = {
+	createdByUserId: number;
+	distance: number;
+	duration: number;
+	geometry: LineStringGeometry;
 	id: number;
 	name: string;
 	pois: {